requirements.txt
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-experimental==0.0.57
openai==1.20.0
pydantic==2.7.0
streamlit==1.33.0
python-dotenv==1.0.0